Caterina bootloader: adjust bootloader delay if A0 LOW
At boot time, configure A0 (PF7) as an input with internal pullup enabled. If the pin reads low (grounded) set the timeout period to 4000 ms instead of the usual 8000 ms. This will be good for testing. If the test works, then we can reduce that to something tiny like 5ms.